# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Moneyball Scout API",
  "description": "Live football data from API-Football, match ratings, form, valuations and manager ratings.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/models/runs": {
   "get": {
    "tags": [
     "system"
    ],
    "summary": "Model Runs",
    "operationId": "model_runs_api_models_runs_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "default": 20,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ModelRunOut"
         },
         "title": "Response Model Runs Api Models Runs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/players": {
   "get": {
    "tags": [
     "players"
    ],
    "summary": "List Players",
    "operationId": "list_players_api_players_get",
    "parameters": [
     {
      "name": "position",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "GK",
          "DEF",
          "MID",
          "FWD"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Position"
      }
     },
     {
      "name": "team_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Team Id"
      }
     },
     {
      "name": "age_max",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Age Max"
      }
     },
     {
      "name": "value_max",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "description": "EUR",
       "title": "Value Max"
      },
      "description": "EUR"
     },
     {
      "name": "min_minutes",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Min Minutes"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "perf_index",
        "form",
        "avg_rating",
        "market_value",
        "minutes",
        "goals",
        "undervalue"
       ],
       "type": "string",
       "default": "perf_index",
       "title": "Sort"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/PlayerRow"
         },
         "title": "Response List Players Api Players Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/players/{player_id}": {
   "get": {
    "tags": [
     "players"
    ],
    "summary": "Player Detail",
    "operationId": "player_detail_api_players__player_id__get",
    "parameters": [
     {
      "name": "player_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Player Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlayerDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/players/{player_id}/ratings": {
   "get": {
    "tags": [
     "players"
    ],
    "summary": "Player Ratings",
    "operationId": "player_ratings_api_players__player_id__ratings_get",
    "parameters": [
     {
      "name": "player_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Player Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/RatingPoint"
         },
         "title": "Response Player Ratings Api Players  Player Id  Ratings Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/players/{player_id}/similar": {
   "get": {
    "tags": [
     "players"
    ],
    "summary": "Similar",
    "operationId": "similar_api_players__player_id__similar_get",
    "parameters": [
     {
      "name": "player_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Player Id"
      }
     },
     {
      "name": "max_value",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Max Value"
      }
     },
     {
      "name": "max_age",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Max Age"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 15,
       "default": 10,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/SimilarRow"
         },
         "title": "Response Similar Api Players  Player Id  Similar Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/moneyball/undervalued": {
   "get": {
    "tags": [
     "players"
    ],
    "summary": "Undervalued",
    "operationId": "undervalued_api_moneyball_undervalued_get",
    "parameters": [
     {
      "name": "position",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "GK",
          "DEF",
          "MID",
          "FWD"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Position"
      }
     },
     {
      "name": "age_max",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Age Max"
      }
     },
     {
      "name": "value_max",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Value Max"
      }
     },
     {
      "name": "min_minutes",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 450,
       "title": "Min Minutes"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "default": 25,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/PlayerRow"
         },
         "title": "Response Undervalued Api Moneyball Undervalued Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/moneyball/scatter": {
   "get": {
    "tags": [
     "players"
    ],
    "summary": "Scatter",
    "operationId": "scatter_api_moneyball_scatter_get",
    "parameters": [
     {
      "name": "position",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "GK",
          "DEF",
          "MID",
          "FWD"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Position"
      }
     },
     {
      "name": "min_minutes",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 270,
       "title": "Min Minutes"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ScatterPoint"
         },
         "title": "Response Scatter Api Moneyball Scatter Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ModelRunOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "git_sha": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Git Sha"
     },
     "data_cutoff": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Data Cutoff"
     },
     "mlflow_run_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mlflow Run Id"
     },
     "metrics": {
      "additionalProperties": true,
      "type": "object",
      "title": "Metrics"
     }
    },
    "type": "object",
    "required": [
     "id",
     "created_at",
     "status",
     "metrics"
    ],
    "title": "ModelRunOut"
   },
   "PlayerDetail": {
    "properties": {
     "player": {
      "$ref": "#/components/schemas/PlayerRow"
     },
     "nationality": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Nationality"
     },
     "birth_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Birth Date"
     },
     "height_cm": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Height Cm"
     },
     "profile_position": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Profile Position"
     },
     "form_sd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Form Sd"
     },
     "per90": {
      "additionalProperties": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ]
      },
      "type": "object",
      "title": "Per90"
     },
     "shrunk_per90": {
      "additionalProperties": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ]
      },
      "type": "object",
      "title": "Shrunk Per90"
     },
     "percentiles": {
      "additionalProperties": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ]
      },
      "type": "object",
      "title": "Percentiles"
     },
     "valuation": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Valuation"
       },
       {
        "type": "null"
       }
      ]
     },
     "value_history": {
      "items": {
       "$ref": "#/components/schemas/ValuePoint"
      },
      "type": "array",
      "title": "Value History"
     },
     "contract_expiration": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contract Expiration"
     }
    },
    "type": "object",
    "required": [
     "player",
     "per90",
     "shrunk_per90",
     "percentiles",
     "value_history"
    ],
    "title": "PlayerDetail"
   },
   "PlayerRow": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "photo": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Photo"
     },
     "team": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/TeamRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "position_group": {
      "type": "string",
      "title": "Position Group"
     },
     "age": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Age"
     },
     "apps": {
      "type": "integer",
      "title": "Apps"
     },
     "minutes": {
      "type": "integer",
      "title": "Minutes"
     },
     "goals": {
      "type": "integer",
      "title": "Goals"
     },
     "assists": {
      "type": "integer",
      "title": "Assists"
     },
     "avg_rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Rating"
     },
     "form": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Form"
     },
     "expected_next": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expected Next"
     },
     "perf_index": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Perf Index"
     },
     "market_value": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Market Value"
     },
     "p50": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "P50"
     },
     "undervalue_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Undervalue Score"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "position_group",
     "apps",
     "minutes",
     "goals",
     "assists"
    ],
    "title": "PlayerRow"
   },
...
   "RatingPoint": {
    "properties": {
     "fixture_id": {
      "type": "integer",
      "title": "Fixture Id"
     },
     "date": {
      "type": "string",
      "format": "date-time",
      "title": "Date"
     },
     "opponent": {
      "$ref": "#/components/schemas/TeamRef"
     },
     "home": {
      "type": "boolean",
      "title": "Home"
     },
     "score": {
      "type": "string",
      "title": "Score"
     },
     "minutes": {
      "type": "integer",
      "title": "Minutes"
     },
     "rating": {
      "type": "number",
      "title": "Rating"
     },
     "api_rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Api Rating"
     },
     "form": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Form"
     },
     "form_lo": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Form Lo"
     },
     "form_hi": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Form Hi"
     },
     "components": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Components"
     }
    },
    "type": "object",
    "required": [
     "fixture_id",
     "date",
     "opponent",
     "home",
     "score",
     "minutes",
     "rating",
     "components"
    ],
    "title": "RatingPoint"
   },
   "ScatterPoint": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "team": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Team"
     },
     "position_group": {
      "type": "string",
      "title": "Position Group"
     },
     "age": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Age"
     },
     "perf_index": {
      "type": "number",
      "title": "Perf Index"
     },
     "market_value": {
      "type": "number",
      "title": "Market Value"
     },
     "p10": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "P10"
     },
     "p50": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "P50"
     },
     "p90": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "P90"
     },
     "undervalue_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Undervalue Score"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "position_group",
     "perf_index",
     "market_value"
    ],
    "title": "ScatterPoint"
   },
   "SimilarRow": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "photo": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Photo"
     },
     "team": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/TeamRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "position_group": {
      "type": "string",
      "title": "Position Group"
     },
     "age": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Age"
     },
     "apps": {
      "type": "integer",
      "title": "Apps"
     },
     "minutes": {
      "type": "integer",
      "title": "Minutes"
     },
     "goals": {
      "type": "integer",
      "title": "Goals"
     },
     "assists": {
      "type": "integer",
      "title": "Assists"
     },
     "avg_rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Rating"
     },
     "form": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Form"
     },
     "expected_next": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expected Next"
     },
     "perf_index": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Perf Index"
     },
     "market_value": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Market Value"
     },
     "p50": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "P50"
     },
     "undervalue_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Undervalue Score"
     },
     "similarity": {
      "type": "number",
      "title": "Similarity"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "position_group",
     "apps",
     "minutes",
     "goals",
     "assists",
     "similarity"
    ],
    "title": "SimilarRow"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "ValuePoint": {
    "properties": {
     "date": {
      "type": "string",
      "format": "date",
      "title": "Date"
     },
     "value": {
      "type": "number",
      "title": "Value"
     }
    },
    "type": "object",
    "required": [
     "date",
     "value"
    ],
    "title": "ValuePoint"
   }
  }
 }
}